fix(apps/netsuite): delegate sync_dimension tasks to workers
