+++ apps/mappings/queue.py
@@ -7,36 +7,6 @@ from apps.mappings.constants import SYNC_METHODS
 from apps.mappings.models import GeneralMapping
 from apps.netsuite.helpers import sync_override_tax_items
 
-
-def get_import_categories_settings(configurations: Configuration):
-    """
-    Get import categories settings
-    :return: is_3d_mapping_enabled, destination_field, destination_sync_methods
-    """
-    destination_sync_methods = []
-    destination_field = None
-    is_3d_mapping_enabled = False
-
-    if configurations.import_items:
-        destination_sync_methods.append(SYNC_METHODS['ITEM'])
-
-    if (configurations.reimbursable_expenses_object and configurations.reimbursable_expenses_object == 'EXPENSE REPORT') or configurations.corporate_credit_card_expenses_object == 'EXPENSE REPORT':
-        destination_sync_methods.append(SYNC_METHODS['EXPENSE_CATEGORY'])
-        destination_field = 'EXPENSE_CATEGORY'
-
-    if configurations.reimbursable_expenses_object != 'EXPENSE REPORT' and (
-        configurations.reimbursable_expenses_object in ('BILL', 'JOURNAL ENTRY')
-        or configurations.corporate_credit_card_expenses_object in ('BILL', 'JOURNAL ENTRY', 'CREDIT CARD CHARGE')):
-        destination_sync_methods.append(SYNC_METHODS['ACCOUNT'])
-        destination_field = 'ACCOUNT'
-
-    if configurations.reimbursable_expenses_object == 'EXPENSE REPORT' and \
-    configurations.corporate_credit_card_expenses_object in ('BILL', 'CREDIT CARD CHARGE', 'JOURNAL ENTRY'):
-        is_3d_mapping_enabled = True
-
-    return is_3d_mapping_enabled, destination_field, destination_sync_methods
-
-
 def construct_tasks_and_chain_import_fields_to_fyle(workspace_id: int):
     """
     Construct tasks and chain import fields to fyle
